Rename folder operation => operations
