+++ src/middleware/aws/operations/dynamodb.rs
