+++ src/middleware/aws/operations/firehose.rs
