+++ src/middleware/aws/operations/sns.rs
